+++ storybook/src/Forms/Text.stories.tsx
@@ -25,6 +25,7 @@
  */
 import React from 'react'
 import {
+  Button,
   FieldText,
   FieldCheckbox,
   Fieldset,
@@ -35,6 +36,8 @@ import {
   Space,
   Text,
   Tooltip,
+  useToggle,
+  SpaceVertical,
 } from '@looker/components'
 
 export const All = () => (
@@ -48,13 +51,13 @@ export const All = () => (
     <Placeholder />
     <Value />
     <ValidationError />
-    <Prefix />
-    <Suffix />
+    <Before />
+    <After />
     <IconBefore />
     <IconAfter />
     <Toggles />
     <AutoResize />
-    <BeforeAfter />
+    <BeforeAfterValidation />
   </Fieldset>
 )
 
@@ -86,8 +89,8 @@ export const ValidationError = () => (
     validationMessage={{ message: 'Error Message', type: 'error' }}
   />
 )
-export const Prefix = () => <FieldText prefix="$" label="Dollars" />
-export const Suffix = () => <FieldText prefix="%" label="Percent" />
+export const Before = () => <FieldText before="$" label="Dollars" />
+export const After = () => <FieldText after="%" label="Percent" />
 export const IconBefore = () => (
   <FieldText iconBefore="GearOutline" label="Settings" />
 )
@@ -141,26 +144,38 @@ export function AutoResize() {
   )
 }
 
-export function BeforeAfter() {
+export function BeforeAfterValidation() {
+  const { value, toggle } = useToggle(true)
+  const validation = value
+    ? { validationMessage: { message: 'Oops!', type: 'error' as 'error' } }
+    : {}
   return (
-    <Space>
-      <FieldText label="iconBefore" iconBefore="Favorite" />
-      <FieldText label="iconAfter" iconAfter="Account" />
-      <FieldText label="prefix" prefix="$" />
-      <FieldText label="suffix" suffix="%" validationType="error" />
-      <FieldText
-        label="before &amp; after"
-        before={
-          <Tooltip content="Some very important info">
-            <Icon name="AddAlerts" size="small" />
-          </Tooltip>
-        }
-        after={
-          <Text fontSize="small" color="critical">
-            Helper text
-          </Text>
-        }
-      />
-    </Space>
+    <SpaceVertical align="start">
+      <Button onClick={toggle}>Toggle error state</Button>
+      <Space>
+        <FieldText label="iconBefore" iconBefore="Favorite" {...validation} />
+        <FieldText label="iconAfter" iconAfter="Account" {...validation} />
+        <FieldText label="before string" before="$" {...validation} />
+        <FieldText label="after string" after="%" {...validation} />
+        <FieldText
+          label="before &amp; after"
+          before={
+            <Tooltip content="Some very important info">
+              <Icon
+                name="AddAlerts"
+                size="small"
+                style={{ cursor: 'default' }}
+              />
+            </Tooltip>
+          }
+          after={
+            <Text fontSize="small" color={value ? 'critical' : 'info'}>
+              Helper text
+            </Text>
+          }
+          {...validation}
+        />
+      </Space>
+    </SpaceVertical>
   )
 }